[Feat] 창고에 장비 파악후 아이템 활성화 기능
스탯과 플레이어 매니저간의 자료구조 동일화까지 진행

+++ Assets/WorkPlace/LCM/Scripts/ScriptableObject/AirTankItem.cs
@@ -7,7 +7,7 @@ public class AirTankItem : ConsumableItem
 {
     [Header("Air Tank Specifics")]
     [Tooltip("이 아이템 사용 시 회복될 산소량입니다.")]
-    public float OxygenRestoreAmount;
+    public double OxygenRestoreAmount;
 
     public override void Use(GameObject user)
     {
